+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "private": true,
   "scripts": {
-    "build": "webpack --mode=production",
-    "dev": "webpack --watch --mode=development",
+    "build": "NODE_ENV=production && gulp",
+    "dev": "NODE_ENV=development && gulp",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -19,13 +19,10 @@
   },
   "homepage": "https://github.com/amyyf/back-to-skull-island#readme",
   "devDependencies": {
-    "css-loader": "^3.2.0",
-    "file-loader": "^4.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "style-loader": "^1.0.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.6",
-    "zip-webpack-plugin": "^3.0.0"
+    "gulp": "^4.0.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-csso": "^3.0.1",
+    "gulp-sourcemaps": "^2.6.5"
   },
   "dependencies": {}
 }
