+++ src/controller.pkg.js
@@ -1,0 +1,71 @@
+// import { init as modelInit, getMapView } from '../model';
+// import { setState as modelSetState, getState, isUpdated } from '../model/state';
+
+const initController = () => {
+  const loop = () => {
+    const state = getState();
+    const heightPadding = Math.ceil(((state.height / tileHeight) - 1) / 2)
+    const widthPadding = Math.ceil(((state.width / tileWidth) - 1) / 2)
+    draw(
+      getMapView(state.position, heightPadding, widthPadding),
+      state
+    ).then(() => isUpdated() && loop());
+  };
+
+  const {
+    getMapView
+  } = initModel();
+
+  const {
+    getState,
+    isUpdated,
+    setState: modelSetState
+  } = initState();
+  const setState = val => modelSetState(val) && loop();
+
+  const onWindowResize = (bodyDimensions) => setState(bodyDimensions);
+  const onKeyDown = (which) => {
+    setState((state) => {
+      let row = 0;
+      let col = 0;
+      switch (which) {
+        case 37:
+          // left
+          col--;
+          break;
+        case 38:
+          // up
+          row--;
+          break;
+        case 39:
+          // right
+          col++;
+          break;
+        case 40:
+          // down
+          row++;
+          break;
+        default:
+          return;
+      }
+      const proposedNewState = {
+        row: state.position.row + row,
+        col: state.position.col + col
+      };
+      const cell = getMapView(proposedNewState, 0, 0);
+      if (cell[0][0] !== 'black') { // is not an invalid cell
+        return { position: proposedNewState};
+      }
+    });
+  }
+  const {
+    draw,
+    tileHeight,
+    tileWidth
+  } = initView(modelSetState, { onWindowResize, onKeyDown });
+
+
+  // init
+  modelSetState({ position: {row: 50, col: 50 }});
+  loop();
+};
