+++ src/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>Back to Skull Island</title>
+    <link rel="stylesheet" href="styles.css">
+  </head>
+  <body>
+  <script type="text/javascript" src="game.min.js"></script></body>
+</html>
