+++ src/index.js
@@ -1,4 +1,1 @@
-import { init } from './controller';
-import './styles.css';
-
-init();
+initController();
