--- webpack.config.js
@@ -1,36 +1,0 @@
-const path = require('path');
-
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ZipPlugin = require('zip-webpack-plugin');
-
-module.exports = {
-  entry: './src/index.js',
-  output: {
-    filename: 'bundle.js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader'
-        ]
-      },
-      {
-        test: /\.(png|gif|jpg|svg)$/,
-        use: [
-          'file-loader'
-        ]
-      }
-    ]
-  },
-  devtool: 'inline-source-map',
-  plugins: [
-    new HtmlWebpackPlugin({
-      title: 'Back to Skull Island'
-    }),
-    new ZipPlugin()
-  ]
-};
